Make all builtin datatypes available from node playground
When detecting the output datatype, node playground has logic to see that the type comes from the generatedTypes.ts file. When detecting this, it would return the available datatypes as the _basicDatatypes_ which did not include the foxglove datatypes. This change fixes the datatype return logic to include all the datatypes that went into building the generatedTypes.ts file.

Fixes: #3046